requirements.txt
sqlalchemy==2.0.21
openai>=0.28.0
langchain==0.2.11
langchain-core==0.2.43
langchain_experimental==0.0.63
langchain-community==0.2.10
langchain-openai==0.1.19
python-decouple==3.8
cachetools==5.4.0
fastapi==0.111.1
jinja2==3.1.4
uvicorn[standard]==0.30.3
websocket-client==1.8.0
sqlean.py==3.45.1
isort
black
